Fix broken num_ticks calculation for hasted dots - it wasn't rounding tick frequency to nearest millisecond.
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@10978 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_action.cpp
@@ -1881,7 +1881,7 @@ int action_t::hasted_num_ticks( timespan_t d ) const
   if ( d < timespan_t::zero )
     d = num_ticks * base_tick_time;
 
-  timespan_t t = timespan_t::from_millis( ( base_tick_time.total_millis() * player_haste ) + 0.5 );
+  timespan_t t = timespan_t::from_millis( (int) ( ( base_tick_time.total_millis() * player_haste ) + 0.5 ) );
 
   double n = d / t;
